Removed unused code, removed comments, implemented missing functions that we will use for update/read/search/delete
