+++ application/classes/Ushahidi/Validator/Form/Contact/Create.php
@@ -10,64 +10,9 @@
  */
 
 
-class Ushahidi_Validator_Form_Contact_Create extends \Ushahidi\Core\Tool\Validator
+class Ushahidi_Validator_Form_Contact_Create extends Ushahidi_Validator_Form_Contact_Update
 {
 
 	protected $default_error_source = 'form_contact';
-
-	protected $form_repo;
-	protected $contact_repo;
-	protected $phone_validator;
-
-	public function setPhoneValidator($validator) {
-		$this->phone_validator = $validator;
-
-	}
-	public function setFormRepo(\Ushahidi\Core\Entity\FormRepository $form_repo)
-	{
-		$this->form_repo = $form_repo;
-	}
-
-	public function setContactRepo(\Ushahidi\Core\Entity\ContactRepository $contact_repo)
-	{
-		$this->contact_repo = $contact_repo;
-	}
-
-//	protected function getRules()
-//	{
-//		return [
-//			'form_id' => [
-//				['digit'],
-//				[[$this->form_repo, 'exists'], [':value']],
-//			],
-//			'role_id' => [
-//				[[$this->role_repo, 'idExists'], [':value']],
-//			],
-//		];
-//	}
-	protected function getRules()
-	{
-		return [
-			'form_id' => [
-				['digit'],
-				[[$this->form_repo, 'exists'], [':value']],
-			],
-			'country_code' => [
-				['not_empty'],
-			],
-			'contact' => [
-				[[$this, 'contactIsValid'], [':value', ':fulldata']],
-			],
-		];
-	}
-	public function contactIsValid($value, $fullData) {
-		try {
-			$number = $this->phone_validator->parse($value, $fullData['country_code']);
-			$isValid = $this->phone_validator->isValidNumber($number);
-			return $isValid;
-		} catch (\libphonenumber\NumberParseException $e) {
-			return false;
-		}
-	}
 
 }
